// src/types.ts
// types.ts
export enum TurkishDialect {
    TURKEY = 1,
    KAZAKH = 2,
    UZBEK = 3,
    AZERBAIJANI = 4,
    UYGHUR = 5,
    BASHKIR = 6,
    TATAR = 7,
    TURKMEN = 8,
    KYRGYZ = 9,
    RUSSIAN = 10,
  }
  
  export enum PersonNameType {
    FEMALE = 1,
    MALE = 2,
    UNISEX = 3,
  }
  
  export enum PersonNameSearchType {
    BY_NAME = 1,
    BY_MEANING = 2,
  }
  
  export enum ScienceArtDict {
    ALL = -1,
    WOODWORK = 0,
    PARASITOLOGY = 1,
    ATHLETICS = 2,
    FOOTBALL = 3,
    LIGHTING = 4,
    COMPUTER_TERMS = 5,
    IT_TERMS = 6,
    BOTANY = 7,
    BIOCHEMISTRY = 8,
    BIOLOGY = 9,
    ETHNOLOGY = 10,
    CRIMINAL_PROCEDURE = 11,
    GYMNASTICS = 12,
    GEOGRAPHY = 13,
    BICYCLE = 14,
    GRAMMAR = 15,
    LINGUISTICS = 16,
    BIOLOGY_ALT = 17,
    PLUMBING = 18,
    LITERATURE = 19,
    EDUCATION = 20,
    ECONOMETRICS = 21,
    PHILOSOPHY = 22,
    PHYSICS = 23,
    PHYSICAL_CHEMISTRY = 24,
    GEOMETRY = 25,
    ASTRONOMY = 26,
    PERFORMING_ARTS = 27,
    GRAMMAR_ALT = 28,
    CUSTOMS = 29,
    WRESTLING = 30,
    FINE_ARTS = 31,
    FOLKLORE = 32,
    NURSING = 33,
    HISTOLOGY = 34,
    ECONOMICS = 35,
    PHARMACY = 36,
    STATISTICS = 37,
    URBAN_PLANNING = 38,
    FENCING = 39,
    CHEMISTRY_2007 = 40,
    CHEMISTRY_1981 = 41,
    LIBRARY_SCIENCE = 42,
    CRIMINOLOGY = 43,
    MINING = 44,
    LOGIC = 45,
    MATHEMATICS = 46,
    CIVIL_LAW = 47,
    METALLURGY = 48,
    NUCLEAR_ENERGY = 49,
    SECONDARY_EDUCATION = 50,
    AUTOMOTIVE = 51,
    PSYCHOLOGY = 52,
    BASKETBALL = 53,
    CINEMA_TV = 54,
    AQUACULTURE = 55,
    AGRICULTURE = 56,
    HISTORY = 57,
    COMMERCE = 58,
    TECHNICAL_TERMS = 59,
    MEDICINE = 60,
    THEATER = 61,
    SOCIOLOGY = 62,
    RULE_TERMS = 63,
    TURKIC_GRAMMAR = 64,
    MEDICAL_TERMS = 65,
    VOLLEYBALL = 66,
    METROLOGY = 67,
    APPLIED_TERMS = 68,
    VETERINARY = 69,
    CONSTRUCTION = 70,
    COPYRIGHT = 71,
    LITERARY_TERMS = 72,
    EARTH_SCIENCE = 73,
    METHODOLOGY = 74,
    BOXING = 75,
    CRAFT = 76,
    ZOOLOGY = 77,
  }
  